backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis / Queue
redis[hiredis]==5.2.1
celery[redis]==5.4.0

# Crawler
aiohttp==3.11.10
playwright==1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0
fake-useragent==1.5.1
tldextract==5.1.3
robotparser==0.0.1

# Data
pydantic==2.10.3
pydantic-settings==2.6.1

# OpenTelemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-instrumentation-aiohttp-client==0.50b0
opentelemetry-instrumentation-celery==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-prometheus==0.50b0
prometheus-client==0.21.1

# Utilities
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==9.0.0
structlog==24.4.0
python-dotenv==1.0.1
